fix: server process won't die
Change the mode of the Flask server to `production` instead of `development`.

The mode of the Falsk server dictates how the server will respond to SIGTERM, SIGKILL,
and other signals. A keyboard interrupt is pretty much the easiest way to kill the
server in `development` mode. Changing the mode to `production` makes it easy to kill
the server with a SIGTERM.