Add a function to asynchronously transmit logs
Add a function to asynchronously transmit logs to the server.

Transmission of log records should be done asynchronously, to prevent blocking the
thread. A durable `aiohttp.ClientSession` object should be used as the session to
prevent recreating the session with each request. Also, a validated URL should be used.
This function was created with these requirements in mind.

Errors can happen during transmission, or when the response is interacted with.
`TransmissionError` and `ResponseError` were created to be raised when these errors are
encountered.

The process of transmission can cause errors. These errors can be classified into
multiple types, so an `enum` was created to classify them easily.

Closes #13.