Transmit logs in a loop
Transmit logs to server asynchronously in a loop.

Logging is a continuous activity. As longs as the application runs, logs will be
generated. Thus, the logs will need to be transmitted to the server continuously,
preferably using a single session. After all the transmissions are done, the status
should be returned.

Closes #28.