Add function to start transmission loop
Create a function to start the `transmission_loop` function asynchronously and block
until the transmission is done.

The `transmission_loop` is asynchronous and must be called from within `asyncio.run`. This will
act as the single point of asynchronous work for the thread it is running on. The `asyncio.run`
function will block until the transmission loop exits, so it is better to encapsulate
this logic within the `start_transmission_loop` function—this way the function can be
run within another thread without blocking other threads.

Closes #30.