Add function to clean up after transmissions (#39)
Add an async function to clean up after the transmission is over.

The set of `asyncio.Task` is used to prevent the garbage collection mechanism from
auto discarding the tasks. The set needs to be cleaned up after each task has completed
execution. Removing a task from the set is an O(1) task. This helps the set from growing
too large and affecting performance.

Closes #34.