Add function to save transmission result (#41)
Add a function to save transmission result in a threadsafe datastructure for append
and pop operations.

Transmission results will be appended from the thread performing transmissions, and
should be readable from the main thread or any other thread. Thus, we need to save the
result of the transmission in a datastructure that can be appended and poped from
opposite ends, while maintaining threadsafety. This will allow us to check the result of
each complete transmission from another thread, since the transmission will occur on a
thread that is seperate from the main thread.

`collections.deque` is a queue implementation that is threadsafe for append and pop from
opposite ends.

Closes #33.