Add a `StructuredMessage` class
Create a `StructuredMessage` class which can be used to capture
messages, and associated context as keyword arguments, and then
serialized to a JSON string.

Structured logging provides context to log messages, so that
a clear and comprehensive picture of the state of the application
can be observed when looking at the logs. This class helps easily
capture log messages, as well as any other associated context, as
a JSON string.

The message will be present in the JSON string under the
key `message`.

Closes #62.