ref:`LogJSONFormatter` for structured logging
Refactor `LogJSONFormatter` to support structured logging, and ensure `exec_info` or
execution information in the log record is stored under a key `exec_info` in the JSON
object string.

Previously, the `LogJSONFormatter` stored the log message string as is. A structured
message will return a JSON encoded string, and storing it as it is means it will have
to be parsed into a separate JSON object to be useful. This creates a rather incomplete
JSON object; the JSON string created leaves it ambiguous what the `message` key value
actually is: another JSON object altogether. Rather than do that, we now encode it as
another JSON object under the `message` key.

The `LogJSONFormatter` previously formatted the exception message and appended it to the
log message string. Now, it stores the exception message under a separate key.

Closes #67.
Closes #57.