fix: prevent duplicate run of GitHub actions
Prevent this GitHub Actions workflow from running after push to `main`.

This workflow runs whenever there is a push to `main` branch of the repository. This
includes after there is a pull request that is merged or squashed. Since we are doing
trunk development, there is no need for a duplicate run.

Closes #20.